fix: Refactores image detection page
Changed API KEY to env variables, set up API for call to GEMINI
