+++ vita/data/datatypes/nutritionalInfo.tsx
@@ -1,0 +1,9 @@
+
+export interface NutritionalInfo {
+    name: string;
+    calories: number;
+    lipids: number;
+    proteins: number;
+    carbohydrates: number;
+    subgroups: string[];
+}
